requirements.txt
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
github3.py==1.3.0
idna==2.10
jwcrypto==0.8
pycparser==2.20
python-dateutil==2.8.1
requests==2.24.0
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.10
